Move webkit_resources.rc to webkit_resources.grd.  Add webkit_resources.vcproj which creates the .rc files and have webkit glue depend on it.
I had to add a flag to grit that allows a resource to be in the
header, but not in the .rc file.  We do this for a test_shell only
resource that is not compiled in.

SCons build also works on windows.

I still need to update the other .sln files and mac build.

Review URL: http://codereview.chromium.org/17466

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bc4451ceed50867267e75b6677c945e39442e332